requirements.txt
feedparser
gtts
google-generativeai
google-cloud-texttospeech
pydub
google-cloud-storage